Refactor for tests
